+++ FlocosEngine/include/tests/test_batch_render.hpp
@@ -20,9 +20,9 @@ private:
   u_short m_GridM;
   u_short m_GridN;
   std::unique_ptr<float[]> m_Positions;
-  size_t m_PosSize;
+  uint m_PosSize;
   std::unique_ptr<uint[]> m_Indices;
-  size_t m_IndSize;
+  uint m_IndSize;
   bool m_Initiated;
   glm::mat4 m_ProjMatrix;
   glm::mat4 m_ViewMatrix;
